# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: MiniSpotify
  contact:
    name: Spotify Support
    url: www.spotify.com
    email: adamwiernicki1991@gmail.com
# during dev, should point to your local machine
host: localhost:3000
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      #parameters:
      #  - name: name
      #    in: query
      #    description: The name of the person to whom to say hello
      #    required: false
      #    type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /playlist:
    x-swager-router-controller:
      hello_world
    post:
      description: Adds new playlist of a given name to database
      operationId: create
      parameters:
        - name: name
          in: query
          description: The name of the new playlist
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      description: display playlist properties and tracks
      operationId: displayPlaylist
      parameters:
        - name: name
          in: query
          description: The name of the called playlist
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/Playlist"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
              
  /swagger:
    x-swagger-pipe: swagger_raw

# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  Playlist:
    type: object
    required:
      - name
      - public
    properties:
      name:
        type: string
      public:
        type: boolean